% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{download_athleticism_plot}
\alias{download_athleticism_plot}
\title{Create Downloadable Plot}
\usage{
download_athleticism_plot(input_df, input_player)
}
\arguments{
\item{input_df}{This is the dataframe}

\item{input_player}{Character string of players}
}
\value{
downloadable plot
}
\description{
Create Downloadable Plot
}
