% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{leaderboard_table}
\alias{leaderboard_table}
\title{Table of player metrics}
\usage{
leaderboard_table(input_df, input_season)
}
\arguments{
\item{input_df}{data frame}

\item{input_season}{selected seasons}
}
\value{
player metric table
}
\description{
Table of player metrics
}
