% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{output_athleticism_plot}
\alias{output_athleticism_plot}
\title{Create RYOE Plot}
\usage{
output_athleticism_plot(input_df, input_player)
}
\arguments{
\item{input_df}{This is the data}

\item{input_player}{Character string of players}
}
\value{
plot
}
\description{
Create RYOE Plot
}
